requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
cryptography==43.0.0
dj-database-url==2.2.0
Django==5.0.7
django-allauth==64.0.0
django-crispy-forms==2.3
django-environ==0.11.2
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.7
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
